+++ app/Http/Controllers/Admin/CategoryController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Models\Category as MainModel;
 use Illuminate\Http\Request;
 use App\Http\Requests\CategoryFormRequest as MainRequest;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 use Config;
 
@@ -38,14 +40,18 @@ class CategoryController extends Controller
     }
     public function index(Request $request)
     {
-        $this->params['filter_status'] = (in_array($request->input('filter_status'),$this->params['status_for_controller'])? $request->input('filter_status','all') : 'all');
-        $this->params['search_field'] = (in_array($request->input('search_field','all'),$this->params['search_field_for_controller'])? $request->input('search_field','all') : 'all');
-        $this->params['search_value'] = $request->input('search_value','');
-        $this->params['filter_display'] = (in_array($request->input('display_field'),$this->params['display_for_controller'])? $request->input('display_field','all') : 'all');
-        $status_counts = $this->model->countByStatus($this->params,['task'=>'admin_count_status']);
-        $items = $this->model->getListItems($this->params, ['task'=>'admin_get_list_items']);
-        // dd($items);
-        return view($this->pathViewName.'index',compact('items','status_counts'))->with("params",$this->params);
+        if(!Gate::allows('category.view')){
+            return redirect('/no-permission');
+        }else{
+            $this->params['filter_status'] = (in_array($request->input('filter_status'),$this->params['status_for_controller'])? $request->input('filter_status','all') : 'all');
+            $this->params['search_field'] = (in_array($request->input('search_field','all'),$this->params['search_field_for_controller'])? $request->input('search_field','all') : 'all');
+            $this->params['search_value'] = $request->input('search_value','');
+            $this->params['filter_display'] = (in_array($request->input('display_field'),$this->params['display_for_controller'])? $request->input('display_field','all') : 'all');
+            $status_counts = $this->model->countByStatus($this->params,['task'=>'admin_count_status']);
+            $items = $this->model->getListItems($this->params, ['task'=>'admin_get_list_items']);
+            // dd($items);
+            return view($this->pathViewName.'index',compact('items','status_counts'))->with("params",$this->params);
+        }    
     }
 
     /**
@@ -53,25 +59,33 @@ class CategoryController extends Controller
      */
     public function changeStatus(Request $request)
     {
-       $id = $request->id;
-       $status = $request->status;
-       if($status =='default'){
-        $status = 'active';
-       }
-       $this->model->updateItem(['id'=> $id,'status'=> $status],['task'=>'admin_change_status']);
-       return redirect()->back()->with('notify', 'Kích hoạt trạng thái thành công');
+        if(!Gate::allows('category.change_status')){
+            return redirect('/no-permission');
+        }else{
+            $id = $request->id;
+            $status = $request->status;
+            if($status =='default'){
+                $status = 'active';
+            }
+            $this->model->updateItem(['id'=> $id,'status'=> $status],['task'=>'admin_change_status']);
+            return redirect()->back()->with('notify', 'Kích hoạt trạng thái thành công');
+        }    
        
     }
     //display in home or not
     public function changeDisplay(Request $request)
     {
-       $id = $request->id;
-       $display = $request->display;
-       if($display =='default'){
-        $display = 'yes';
-       }
-       $this->model->updateItem(['id'=> $id,'display'=> $display],['task'=>'admin_change_display']);
-       return redirect()->back()->with('notify', 'Kích hoạt trạng thái hiển thị thành công');
+        if(!Gate::allows('category.change_display')){
+            return redirect('/no-permission');
+        }else{
+            $id = $request->id;
+            $display = $request->display;
+            if($display =='default'){
+                $display = 'yes';
+            }
+            $this->model->updateItem(['id'=> $id,'display'=> $display],['task'=>'admin_change_display']);
+            return redirect()->back()->with('notify', 'Kích hoạt trạng thái hiển thị thành công');
+        }    
        
     }
 
@@ -80,15 +94,23 @@ class CategoryController extends Controller
      */
     public function showFormAdd(Request $request)
     {
-        return view($this->pathViewName.'form_add');
+        if(!Gate::allows('category.add')){
+            return redirect('/no-permission');
+        }else{
+            return view($this->pathViewName.'form_add');
+        }    
     }
 
     public function showFormEdit(Request $request)
     {
-        $id = $request->id;
-        $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id)
-        // dd($item);
-        return view($this->pathViewName.'form_edit', compact('item'));
+        if(!Gate::allows('category.edit')){
+            return redirect('/no-permission');
+        }else{
+            $id = $request->id;
+            $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id)
+            // dd($item);
+            return view($this->pathViewName.'form_edit', compact('item'));
+        }
     }
 
     /**
@@ -96,19 +118,23 @@ class CategoryController extends Controller
      */
     public function save(MainRequest $request)
     {
-        if($request->method() == 'POST'){
-            $params = $request->all();        
-            // dd($params);    
-            if(!empty($request->id)){
-                $this->model->updateItem($params,['task'=>'admin_edit_item']);
-                $notify = "Chỉnh Sửa Thông Tin Thành Công";
-            }else{
-                $params['thumb'] = $request->file('thumb');
-                $this->model->insertItem($params,['task'=>'admin_add_new_item']);
-                $notify = "Thêm Mới Thành Công";
+        if(!Gate::allows('category.save')){
+            return redirect('/no-permission');
+        }else{
+            if($request->method() == 'POST'){
+                $params = $request->all();        
+                // dd($params);    
+                if(!empty($request->id)){
+                    $this->model->updateItem($params,['task'=>'admin_edit_item']);
+                    $notify = "Chỉnh Sửa Thông Tin Thành Công";
+                }else{
+                    $params['thumb'] = $request->file('thumb');
+                    $this->model->insertItem($params,['task'=>'admin_add_new_item']);
+                    $notify = "Thêm Mới Thành Công";
+                }
+                return redirect()->route($this->controllerName.'.index')->with('notify', $notify);
             }
-            return redirect()->route($this->controllerName.'.index')->with('notify', $notify);
-        }
+        }    
     }
 
     /**
@@ -116,11 +142,15 @@ class CategoryController extends Controller
      */
     public function delete(Request $request)
     {
-        $id = $request->id;
-        $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id) infos in the db
-        $this->model->deleteItem($item,['task'=>'admin_delete_item']);
-        $notify = "Xóa Thành Công";
-        return redirect()->route($this->controllerName.'.index')->with('notify', $notify);
+        if(!Gate::allows('category.delete')){
+            return redirect('/no-permission');
+        }else{
+            $id = $request->id;
+            $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id) infos in the db
+            $this->model->deleteItem($item,['task'=>'admin_delete_item']);
+            $notify = "Xóa Thành Công";
+            return redirect()->route($this->controllerName.'.index')->with('notify', $notify);
+        }    
     }
 
 }
