+++ app/Http/Controllers/Admin/CountryController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Models\Country as MainModel;
 use Illuminate\Http\Request;
 use App\Http\Requests\CountryFormRequest as MainRequest;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 use Config;
 
@@ -38,6 +40,9 @@ class CountryController extends Controller
     }
     public function index(Request $request)
     {
+        if(!Gate::allows('country.view')){
+            return redirect('/no-permission');
+        }
         $this->params['filter_status'] = (in_array($request->input('filter_status'),$this->params['status_for_controller'])? $request->input('filter_status','all') : 'all');
         $this->params['search_field'] = (in_array($request->input('search_field','all'),$this->params['search_field_for_controller'])? $request->input('search_field','all') : 'all');
         $this->params['search_value'] = $request->input('search_value','');
@@ -45,7 +50,7 @@ class CountryController extends Controller
         $status_counts = $this->model->countByStatus($this->params,['task'=>'admin_count_status']);
         $items = $this->model->getListItems($this->params, ['task'=>'admin_get_list_items']);
         // dd($items);
-        return view($this->pathViewName.'index',compact('items','status_counts'))->with("params",$this->params);
+        return view($this->pathViewName.'index',compact('items','status_counts'))->with("params",$this->params);   
     }
 
     /**
@@ -53,25 +58,30 @@ class CountryController extends Controller
      */
     public function changeStatus(Request $request)
     {
-       $id = $request->id;
-       $status = $request->status;
-       if($status =='default'){
-        $status = 'active';
-       }
-       $this->model->updateItem(['id'=> $id,'status'=> $status],['task'=>'admin_change_status']);
-       return redirect()->back()->with('notify', 'Kích hoạt trạng thái thành công');
-       
+        if(!Gate::allows('country.change_status')){
+            return redirect('/no-permission');
+        }
+        $id = $request->id;
+        $status = $request->status;
+        if($status =='default'){
+            $status = 'active';
+        }
+        $this->model->updateItem(['id'=> $id,'status'=> $status],['task'=>'admin_change_status']);
+        return redirect()->back()->with('notify', 'Kích hoạt trạng thái thành công');
     }
     //display in home or not
     public function changeDisplay(Request $request)
     {
-       $id = $request->id;
-       $display = $request->display;
-       if($display =='default'){
-        $display = 'yes';
-       }
-       $this->model->updateItem(['id'=> $id,'display'=> $display],['task'=>'admin_change_display']);
-       return redirect()->back()->with('notify', 'Kích hoạt trạng thái hiển thị thành công');
+        if(!Gate::allows('country.change_display')){
+            return redirect('/no-permission');
+        }
+        $id = $request->id;
+        $display = $request->display;
+        if($display =='default'){
+            $display = 'yes';
+        }
+        $this->model->updateItem(['id'=> $id,'display'=> $display],['task'=>'admin_change_display']);
+        return redirect()->back()->with('notify', 'Kích hoạt trạng thái hiển thị thành công');
        
     }
 
@@ -80,11 +90,17 @@ class CountryController extends Controller
      */
     public function showFormAdd(Request $request)
     {
+        if(!Gate::allows('country.add')){
+            return redirect('/no-permission');
+        }
         return view($this->pathViewName.'form_add');
     }
 
     public function showFormEdit(Request $request)
     {
+        if(!Gate::allows('country.edit')){
+            return redirect('/no-permission');
+        }
         $id = $request->id;
         $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id)
         // dd($item);
@@ -96,6 +112,9 @@ class CountryController extends Controller
      */
     public function save(MainRequest $request)
     {
+        if(!Gate::allows('country.save')){
+            return redirect('/no-permission');
+        }
         if($request->method() == 'POST'){
             $params = $request->all();        
             // dd($params);    
@@ -114,6 +133,9 @@ class CountryController extends Controller
      */
     public function delete(Request $request)
     {
+        if(!Gate::allows(['country.delete'])){
+            return redirect('/no-permission');
+        }
         $id = $request->id;
         $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id) infos in the db
         $this->model->deleteItem($item,['task'=>'admin_delete_item']);
