+++ app/Http/Controllers/Admin/SliderController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Models\Slider as MainModel;
 use Illuminate\Http\Request;
 use App\Http\Requests\SliderFormRequest as MainRequest;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 use Config;
 
@@ -38,6 +40,9 @@ class SliderController extends Controller
     }
     public function index(Request $request)
     {
+        if(!Gate::allows('slider.view')){
+            return redirect('/no-permission');
+        }
         $this->params['filter_status'] = (in_array($request->input('filter_status'),$this->params['status_for_controller'])? $request->input('filter_status','all') : 'all');
         $this->params['search_field'] = (in_array($request->input('search_field','all'),$this->params['search_field_for_controller'])? $request->input('search_field','all') : 'all');
         $this->params['search_value'] = $request->input('search_value','');
@@ -53,25 +58,31 @@ class SliderController extends Controller
      */
     public function changeStatus(Request $request)
     {
-       $id = $request->id;
-       $status = $request->status;
-       if($status =='default'){
-        $status = 'active';
-       }
-       $this->model->updateItem(['id'=> $id,'status'=> $status],['task'=>'admin_change_status']);
-       return redirect()->back()->with('notify', 'Kích hoạt trạng thái thành công');
+        if(!Gate::allows('slider.change_status')){
+            return redirect('/no-permission');
+        }
+        $id = $request->id;
+        $status = $request->status;
+        if($status =='default'){
+            $status = 'active';
+        }
+        $this->model->updateItem(['id'=> $id,'status'=> $status],['task'=>'admin_change_status']);
+        return redirect()->back()->with('notify', 'Kích hoạt trạng thái thành công');
        
     }
     //display in home or not
     public function changeDisplay(Request $request)
     {
-       $id = $request->id;
-       $display = $request->display;
-       if($display =='default'){
-        $display = 'yes';
-       }
-       $this->model->updateItem(['id'=> $id,'display'=> $display],['task'=>'admin_change_display']);
-       return redirect()->back()->with('notify', 'Kích hoạt trạng thái hiển thị thành công');
+        if(!Gate::allows('slider.change_display')){
+            return redirect('/no-permission');
+        }
+        $id = $request->id;
+        $display = $request->display;
+        if($display =='default'){
+            $display = 'yes';
+        }
+        $this->model->updateItem(['id'=> $id,'display'=> $display],['task'=>'admin_change_display']);
+        return redirect()->back()->with('notify', 'Kích hoạt trạng thái hiển thị thành công');
        
     }
 
@@ -80,11 +91,17 @@ class SliderController extends Controller
      */
     public function showFormAdd(Request $request)
     {
+        if(!Gate::allows('slider.add')){
+            return redirect('/no-permission');
+        }
         return view($this->pathViewName.'form_add');
     }
 
     public function showFormEdit(Request $request)
     {
+        if(!Gate::allows('slider.edit')){
+            return redirect('/no-permission');
+        }
         $id = $request->id;
         $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id)
         // dd($item);
@@ -96,6 +113,9 @@ class SliderController extends Controller
      */
     public function save(MainRequest $request)
     {
+        if(!Gate::allows('slider.save')){
+            return redirect('/no-permission');
+        }
         if($request->method() == 'POST'){
             $params = $request->all();            
             if(!empty($request->id)){
@@ -115,6 +135,9 @@ class SliderController extends Controller
      */
     public function delete(Request $request)
     {
+        if(!Gate::allows('slider.delete')){
+            return redirect('/no-permission');
+        }
         $id = $request->id;
         $item = $this->model->getItem($id,['task'=>'admin_get_item']); //find($id) infos in the db
         $this->model->deleteItem($item,['task'=>'admin_delete_item']);
@@ -122,19 +145,4 @@ class SliderController extends Controller
         return redirect()->route($this->controllerName.'.index')->with('notify', $notify);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, Slider $slider)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Slider $slider)
-    {
-        //
-    }
 }
